+++ crates/model/src/orders/limit.rs
@@ -133,7 +133,7 @@ impl LimitOrder {
         Ok(Self {
             core: OrderCore::new(init_order),
             price,
-            expire_time: expire_time.or(Some(UnixNanos::default())),
+            expire_time,
             is_post_only: post_only,
             display_qty,
             trigger_instrument_id,